Rename slice for student main
