+++ src/store/index.js
@@ -1,5 +1,4 @@
 import { createStore, applyMiddleware, compose } from 'redux';
-
 import ReduxThunk from 'redux-thunk';
 import reducer from './rootReducer';
